Added "/" to match file check
If two tables exist where the name of one is a substring of the other (e.g. `table1` and `table1_duplicate`) then linting for the first table will pick up files for the second. This commit adds a "/" to the startswith clause that gets the matching files for each table, so that this doesn't happen.